backend/pyproject.toml
[tool.poetry]
name = "vinhos.com"
version = "0.1.0"
description = "Wine add website"
authors = ["Roberto Caiola <robccsilva@gmail.com>, Sérgio den Boer <den.boer.sergio@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
flask-restful = "^0.3.9"
flask-sqlalchemy = "^3.0.2"
flasgger = "^0.9.5"
uwsgi = "^2.0.21"
pymysql = "^1.0.2"
flask-migrate = "^4.0.1"
pytest = "^7.2.1"
flask-jwt-extended = "^4.4.4"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
flake8 = "^6.0.0"
pytest = "^7.2.0"
safety = "^2.3.5"
pytest-flask = "^1.2.0"
pytest-faker = "^2.0.0"
factory-boy = "^3.2.1"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
